+++ crates/client/db/src/tests/common/mod.rs
@@ -17,11 +17,9 @@ use starknet_types_core::felt::Felt;
 pub mod temp_db {
     use crate::DatabaseService;
     use mp_chain_config::ChainConfig;
-    use tempfile::TempDir;
     pub async fn temp_db() -> DatabaseService {
-        let temp_dir = TempDir::new().unwrap();
         let chain_config = std::sync::Arc::new(ChainConfig::madara_devnet());
-        DatabaseService::new(temp_dir.path(), None, false, chain_config).await.unwrap()
+        DatabaseService::open_for_testing(chain_config)
     }
 }