+++ crates/client/devnet/src/lib.rs
@@ -189,6 +189,7 @@ mod tests {
     use mc_mempool::block_production::BlockProductionTask;
     use mc_mempool::MempoolProvider;
     use mc_mempool::{transaction_hash, L1DataProvider, Mempool, MockL1DataProvider};
+    use mc_metrics::MetricsRegistry;
     use mp_block::header::L1DataAvailabilityMode;
     use mp_block::{BlockId, BlockTag};
     use mp_class::ClassInfo;
@@ -209,7 +210,7 @@ mod tests {
     struct DevnetForTesting {
         backend: Arc<MadaraBackend>,
         contracts: DevnetKeys,
-        block_production: BlockProductionTask,
+        block_production: BlockProductionTask<Mempool>,
         mempool: Arc<Mempool>,
     }
 
@@ -299,7 +300,8 @@ mod tests {
         let chain_config = Arc::new(ChainConfig::madara_devnet());
         let block = g.build(&chain_config).unwrap();
         let backend = MadaraBackend::open_for_testing(Arc::clone(&chain_config));
-        let importer = Arc::new(BlockImporter::new(Arc::clone(&backend)));
+        let importer =
+            Arc::new(BlockImporter::new(Arc::clone(&backend), &MetricsRegistry::dummy(), None, true).unwrap());
 
         println!("{:?}", block.state_diff);
         tokio::runtime::Runtime::new()